[FIX] #43 하위 카테고리들 상위 카테고리로 매핑
