+++ savit-client/src/utils/category.ts
@@ -1,0 +1,33 @@
+/**
+ * 카테고리를 대분류로 매핑하는 함수
+ */
+export const mapCategoryToMainCategory = (category: string): '식비' | '교통' | '생활' | '문화' | '기타' => {
+  const categoryMapping: { [key: string]: '식비' | '교통' | '생활' | '문화' | '기타' } = {
+    // 식비 관련
+    '식당': '식비',
+    '카페': '식비',
+    '배달': '식비',
+    
+    // 교통 관련
+    '대중교통': '교통',
+    '택시': '교통',
+    
+    // 생활 관련
+    '통신비': '생활',
+    '공과금': '생활',
+    '편의점/마트': '생활',
+    '의료비': '생활',
+    '교육': '생활',
+    
+    // 문화 관련
+    '공연': '문화',
+    '쇼핑': '문화',
+    '유흥': '문화',
+    '영화': '문화',
+    '정기구독': '문화',
+    
+    // 기본값은 기타
+  }
+  
+  return categoryMapping[category] || '기타'
+}
